% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_files.R
\name{add_files}
\alias{add_files}
\title{Add filepath information}
\usage{
add_files(
  design,
  directory = "./",
  prefix = NA,
  prefix_by = NA,
  suffix = NA,
  sufix_by = NA,
  extension = ""
)
}
\arguments{
\item{design}{Experiment Design}

\item{directory}{Directory the files will be held in}

\item{prefix}{Prefix to give all files}

\item{prefix_by}{Prefix by a particular condition}

\item{suffix}{Suffix to give all files}

\item{sufix_by}{Suffix by a particular condition}

\item{extension}{File extension of files}
}
\value{

}
\description{
Add filepath information
}
