% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_manip.R
\name{add_manipulations}
\alias{add_manipulations}
\title{Add Experimental Manipulations}
\usage{
add_manipulations(design, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Add Experimental Manipulations
}
