% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stimuli.R
\name{add_stimuli}
\alias{add_stimuli}
\title{Add stimuli that don't vary by manipulation}
\usage{
add_stimuli(design, ...)
}
\arguments{
\item{...}{Series of unquoted stimuli. If you want to pass a (character
vector) variable, you should unquote it with !!.}
}
\value{

}
\description{
add_stimuli(stim) will do the same thing as add_stimuli_by(~stim). If you
have a variable such as stim <- c('one','two') you should pass this as
add_stimuli(!!stim). You may use add_stimuli(c('one','two')), but this
requires a few more key strokes than add_stimuli(one, two).
}
\examples{

# All acceptable ways to add audio and image variables (names arbitrary)
test_design <- new_design() \%>\% add_items(critical = 2, filler=1)
stims <- c('audio','image')
test_design \%>\% add_stimuli(audio, image)
test_design \%>\% add_stimuli('audio', 'image')
test_design \%>\% add_stimuli(c('audio', 'image'))
test_design \%>\% add_stimuli('audio + image')
test_design \%>\% add_stimuli(!!stims)
test_design \%>\% add_stimuli_by(~audio + image)

}
