% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stimulus_list.R
\name{create_stimulus_list}
\alias{create_stimulus_list}
\title{Create stimulus list}
\usage{
create_stimulus_list(design)
}
\arguments{
\item{design}{Experiment design}
}
\description{
Create stimulus list
}
