% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Create Design conditions}
\usage{
create_template(design)
}
\arguments{
\item{design}{}
}
\value{
A data frame
}
\description{
Create Design conditions
}
