% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_presentation.R
\name{.add_stimulus_presentation}
\alias{.add_stimulus_presentation}
\title{Create stimulus presentation}
\usage{
.add_stimulus_presentation(
  design,
  formula_manipulation,
  formula_columns,
  manipulation_stimuli,
  is_crossed,
  is_ordered
)
}
\arguments{
\item{design}{Stimulist object}

\item{formula_manipulation}{Vector of strings of manipulation names}

\item{formula_columns}{Columns to add}

\item{manipulation_stimuli}{Expanded grid of stimuli}

\item{is_crossed}{Boolean if should be crossed presentation}

\item{is_ordered}{Boolean if uncrossed presentation has ordering}
}
\description{
Create stimulus presentation
}
