% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_formulas.R
\name{.clean_formulas}
\alias{.clean_formulas}
\title{Clean formulas}
\usage{
.clean_formulas(formulas)
}
\arguments{
\item{formulas}{List of formulas}
}
\description{
Helper to merge formulas that should have been written together as one
}
