% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_manipulations.R
\name{.cross_manipulations}
\alias{.cross_manipulations}
\title{Cross manipulations}
\usage{
.cross_manipulations(design, manipulations)
}
\arguments{
\item{design}{Stimulist object}

\item{manipulations}{Vector of strings of manipulation names}
}
\description{
Helper to cross the levels of two or more manipulations that affect the value
on a trial
}
