% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_table.R
\name{.get_label_table}
\alias{.get_label_table}
\title{Get table of item labels}
\usage{
.get_label_table(design)
}
\arguments{
\item{design}{}
}
\description{
This is used to create a table of labels and their respective items to left
join in to the final table
}
