% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_presentation.R
\name{.make_crossed_presentation}
\alias{.make_crossed_presentation}
\title{Create presentation of crossed manipulations}
\usage{
.make_crossed_presentation(design, manipulations, columns)
}
\arguments{
\item{design}{Stimulist object}

\item{manipulations}{Vector of strings of manipulation names}

\item{columns}{Columns to add}
}
\description{
Helper to make a crossed presentation of stimuli
}
