% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_presentation.R
\name{.make_presentation}
\alias{.make_presentation}
\title{Create presentation of uncrossed manipulations}
\usage{
.make_presentation(design, manipulation, columns)
}
\arguments{
\item{design}{Stimulist object}

\item{manipulation}{Manipulation names}

\item{columns}{Columns to add}
}
\description{
Helper to make a presentation of stimuli
}
