% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus_table.R
\name{get_stim_table}
\alias{get_stim_table}
\title{Create table for stimuli}
\usage{
get_stim_table(design, add_id = T)
}
\arguments{
\item{add_id}{}
}
\value{

}
\description{
Create table for stimuli
}
