% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_template.R
\name{merge_template}
\alias{merge_template}
\title{Merge settings from template}
\usage{
merge_template(design, template_path)
}
\arguments{
\item{template_path}{}
}
\value{

}
\description{
After using save_stimulus_template() to create a blank excel workbook, you
can then fill it out with whatever file names or text you need. But,
you'll probably want to load that back into the design you made so you can
save counterbalanced lists of what you created. This function will fill in
any empty columns in the complete_experiment table with what you put into
the blank template file. The function will automatically handle cases where
you have permutations of orders.
}
