% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_n_of.R
\name{present_n_of}
\alias{present_n_of}
\title{Present n of the conditions of a manipulation}
\usage{
present_n_of(design, manipulation, n)
}
\arguments{
\item{design}{Experimental design}

\item{manipulation}{Manipulation, can be quoted or unquoted}

\item{n}{Number to present}
}
\description{
Present n of the conditions of a manipulation
}
