% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_columns.R
\name{relocate_columns}
\alias{relocate_columns}
\title{Relocate columns in completed experiment}
\usage{
relocate_columns(design, ...)
}
\arguments{
\item{design}{Experiment design}

\item{...}{Series of unquoted columns to move to the far left}
}
\description{
This is just a wrapper around dplyr::relocate to manipulate the ordering
of columns in the complete_experiment entry. Right now it doesn't allow for
the use of .before and .after, but I'll fix that later.
}
