% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_json.R
\name{save_json}
\alias{save_json}
\title{Save JSON files}
\usage{
save_json(
  design,
  filename = gsub(" ", "_", design[["name"]]),
  objectname = "stimulist",
  separate_items = FALSE,
  as_one_file = FALSE
)
}
\arguments{
\item{design}{Experimental design}

\item{filename}{Filename you want to use for each list (number will be appended)}

\item{objectname}{Name you want to use for the js object in the file}

\item{separate_items}{Should different item types be saved in separate files}

\item{as_one_file}{Should all stimuli be saved as one file (NOT recommended)}
}
\description{
This function does the same as save_lists, but in json format instead of
csv format. It will write .js files that you can import directly into your
experiment. Stimuli that present multiple levels on a single trial will be
saved as an object with the stimulus name. For example, audio_1  = x
and audio_2 = y will be converted into audio = {audio_1 = x, audio_2 = y}.
If x and y are greater than length 1 for some reason, they will be converted
to an an array within the audio object.
}
