% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lists.R
\name{save_lists}
\alias{save_lists}
\title{Save stimulus lists}
\usage{
save_lists(
  design,
  filename = design[["name"]],
  path = getwd(),
  separate_items = FALSE,
  as_one_file = FALSE,
  remove_type = FALSE
)
}
\arguments{
\item{design}{Experiment design}

\item{filename}{Name to use for output files}

\item{path}{Directory to save files into (default is getwd())}

\item{separate_items}{Whether to save separate files for different types of
items (eg critical and filler lists are saved separately)}

\item{as_one_file}{Whether you want to save all trials as 1 file}

\item{remove_type}{Should the type column be removed from the output file? If
TRUE and there's only 1 type, the item column is renamed to the 1 value of the
type column. If FALSE, type column is retained. If a string, will remove the
type column and rename the item column to the given string.}
}
\description{
Given a counterbalanced experiment from fill_experiment(), saves the individual
counterbalanced lists. In the future there will be an option for JSON files
or csv files, but for now it does csv files.
}
