% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_filepaths.R
\name{set_filepaths}
\alias{set_filepaths}
\title{Set filepaths on template}
\usage{
set_filepaths(design, template, name_by = "stimulus")
}
\arguments{
\item{name_by}{}
}
\value{

}
\description{
Set filepaths on template
}
