% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lists_by.R
\name{split_lists_by}
\alias{split_lists_by}
\title{Split lists by}
\usage{
split_lists_by(design, ...)
}
\arguments{
\item{design}{Experiment design}

\item{...}{a series of unquoted manipulation or variable that you
want to split the lists by. For example, if you have the same set up but in
3 different languages, you may opt to split the lists by language. Right now
it will overwrite any counterbalancing that's set in the output files.}
}
\description{
Split lists by
}
