% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_manip.R
\name{add_manipulations}
\alias{add_manipulations}
\title{Add Experimental Manipulations}
\usage{
add_manipulations(design, ...)
}
\arguments{
\item{design}{Experiment design}

\item{...}{A series of named arguments with manipulation levels}
}
\description{
Add Experimental Manipulations
}
