% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stimuli_by.R
\name{add_stimuli_by}
\alias{add_stimuli_by}
\title{Add Stimuli by Manipulations}
\usage{
add_stimuli_by(design, ...)
}
\arguments{
\item{design}{Design}

\item{...}{A series of formulas where the left hand side is a manipulation
and the right hand side is a series of variable names}
}
\value{

}
\description{
Add Stimuli by Manipulations
}
\examples{

new_design() \%>\%
add_manipulations(context = c("IN", "OUT"), contour = c("HLL", "LLL")) \%>\%
add_trials(critical = 12, filler = 8) \%>\%
add_stimuli_by(context ~ preamble_text + question_text, contour ~ audio_file)

}
