% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimulus_table.R
\name{get_stim_table}
\alias{get_stim_table}
\title{Create table for stimuli}
\usage{
get_stim_table(design, add_id = TRUE)
}
\arguments{
\item{design}{Experiment design}

\item{add_id}{Whether to add an id to join by}
}
\description{
Create table for stimuli
}
