% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_filenames_by.R
\name{glue_filenames_by}
\alias{glue_filenames_by}
\title{Glue filenames by glue string}
\usage{
glue_filenames_by(design, ..., as_levels = FALSE)
}
\arguments{
\item{design}{Experiment design}

\item{...}{A list of formulas mapping stimuli to glue-formatted strings}

\item{as_levels}{Whether to use the labels or integers representing each level}
}
\value{

}
\description{
Requires the {glue} package. Similar syntax to add_stimuli_by(), but here
you're specifying a glue string for a particular stimulus. Glue strings
utilize {brackets} to evaluate an expression within a string. In this case,
you'll be replacing a label in brackets with its value in the
completed_experiment data frame. Eg:
}
\details{
... %>%
add_stimuli_by(manip1*manip2 ~ audio_file) %>%
fill_experiment() %>%
glue_filenames_by(audio_file ~ "sounds/{manip1}_{manip2}.wav)

This must be run after using fill_experiment.
}
