% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_lists.R
\name{save_lists}
\alias{save_lists}
\title{Save stimulus lists}
\usage{
save_lists(
  design,
  filename = "experiment",
  path = getwd(),
  separate_items = FALSE,
  as_one_file = FALSE
)
}
\arguments{
\item{separate_items}{}
}
\value{

}
\description{
Given a counterbalanced experiment from fill_experiment(), saves the individual
counterbalanced lists. In the future there will be an option for JSON files
or csv files, but for now it does csv files.
}
