% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_stimuli_templates.R
\name{save_stimuli_templates}
\alias{save_stimuli_templates}
\title{Save stimuli templates}
\usage{
save_stimuli_templates(design, as_workbook = TRUE)
}
\arguments{
\item{design}{Experiment design}

\item{as_workbook}{Whether to save as an excel workbook or not}
}
\value{

}
\description{
After adding stimuli with add_stimuli_by(), this will save .csv files for
you to fill in in excel. The completed files will be read in later to merge
into one large table that can be counterbalanced into different lists.
}
